refactor(builder): extrahiere methodology.rs aus single/mod.rs
- build_appendix_block_from_normalized, build_methodology,
  build_confidence_summary, build_capability_matrix → single/methodology.rs
